+++ CHANGES.md
@@ -30,6 +30,7 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
   (or use a custom warning handler).
 - Added the `responseBoolean` utility function,
   to get a boolean out of a response value regardless of `formatversion`.
+- Updated dependencies.
 
 ## v0.3.0 (2021-10-10)